+++ examples/spin-redis/app.py
@@ -1,8 +1,8 @@
-from spin_sdk.http import simple
-from spin_sdk.http.simple import Request, Response
+from spin_sdk.http import IncomingHandler
+from spin_sdk.http import Request, Response
 from spin_sdk import redis
 
-class IncomingHandler(simple.IncomingHandler):
+class IncomingHandler(IncomingHandler):
     def handle_request(self, request: Request) -> Response:
         with redis.open("redis://localhost:6379") as db:
             print(db.get("test"))
